requirements.txt
Flask==3.1.2
flask-cors==4.0.0
pandas==2.1.0
scikit-learn==1.3.0
numpy==1.24.0
python-dotenv==1.2.1
antrhopic==0.73.0